+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Include user selected filters in pagination links
   ([#204](https://github.com/Tubeshade/Tubeshade/issues/204))
+- Remove ignored status from video after download
+  ([#200](https://github.com/Tubeshade/Tubeshade/issues/200))
 
 ## [0.1.0] - 2025-12-06